#20 [Add] 더미 모델 추가
